integrate into existing users api
