+++ api/router.go
@@ -47,5 +47,4 @@ func ConfigGinRouter(router *gin.Engine) {
 	configGinBookmarksRouter(router, daoWrapper)
 	configMaintenanceRouter(router, daoWrapper)
 	configSemestersRouter(router, daoWrapper)
-	configUserRouter(router, daoWrapper)
 }
