+++ api/users.go
@@ -27,6 +27,8 @@ func configGinUsersRouter(router *gin.Engine, daoWrapper dao.DaoWrapper) {
 	router.POST("/api/users/settings/greeting", routes.updatePreferredGreeting)
 	router.POST("/api/users/settings/playbackSpeeds", routes.updatePlaybackSpeeds)
 
+	router.POST("/api/users/resetPassword", routes.resetPassword)
+
 	courses := router.Group("/api/users/courses")
 	{
 		courses.GET("/:id/pin", routes.getPinForCourse)
@@ -711,6 +713,49 @@ func (r usersRoutes) exportPersonalData(c *gin.Context) {
 		return
 	}
 	_, _ = c.Writer.Write(marshal)
+}
+
+func (r usersRoutes) resetPassword(c *gin.Context) {
+	type resetPasswordRequest struct {
+		Username string `json:"username"`
+	}
+	var req resetPasswordRequest
+	err := c.BindJSON(&req)
+	if err != nil {
+		_ = c.AbortWithError(http.StatusBadRequest, tools.RequestError{
+			Status:        http.StatusBadRequest,
+			CustomMessage: "Can't bind request body",
+			Err:           err,
+		})
+		return
+	}
+
+	// continue in goroutine to prevent timing attacks
+	go func() {
+		user, err := r.UsersDao.GetUserByEmail(c, req.Username)
+		if err != nil && errors.Is(err, gorm.ErrRecordNotFound) {
+			// wrong username/email -> pass
+			return
+		}
+		if err != nil {
+			log.WithError(err).Error("can't get user for password reset")
+			return
+		}
+		link, err := r.UsersDao.CreateRegisterLink(c, user)
+		if err != nil {
+			log.WithError(err).Error("can't create register link")
+			return
+		}
+		err = r.EmailDao.Create(c, &model.Email{
+			From:    tools.Cfg.Mail.Sender,
+			To:      user.Email.String,
+			Subject: "TUM-Live: Reset Password",
+			Body:    "Hi! \n\nYou can reset your TUM-Live password by clicking on the following link: \n\n" + tools.Cfg.WebUrl + "/setPassword/" + link.RegisterSecret + "\n\nIf you did not request a password reset, please ignore this email. \n\nBest regards",
+		})
+		if err != nil {
+			log.WithError(err).Error("can't save reset password email")
+		}
+	}()
 }
 
 type personalData struct {
